requirements.txt
paho-mqtt==1.6.1
pydantic==2.4.2
python-dotenv==1.0.0
